+++ src/campaigns/CampaignListItem.mdx
@@ -1,0 +1,13 @@
+---
+name: CampaignListItem
+menu: Components
+---
+
+import CampaignListItem from './CampaignListItem'
+import { Playground } from 'docz'
+
+# Einzeleintrag der Kampagenliste
+
+<Playground>
+  <CampaignListItem />
+</Playground>
