+++ src/campaigns/CampaignsList.mdx
@@ -1,0 +1,13 @@
+---
+name: CampaignsList
+menu: Components
+---
+
+import CampaignsList from './CampaignsList'
+import { Playground } from 'docz'
+
+# Auflistung aller Kampagnen
+
+<Playground>
+  <CampaignsList />
+</Playground>
