+++ src/common/BrowseBack.mdx
@@ -1,0 +1,13 @@
+---
+name: BrowseBack
+menu: Components
+---
+
+import BrowseBack from './BrowseBack'
+import { Playground } from 'docz'
+
+# Footer
+
+<Playground>
+  <BrowseBack />
+</Playground>
